Refactor contract page layout and remove unused bien_ban files. Simplified text widget usage and improved code readability. Added new routes for inventory reconciliation and stocktaking reports.
